pyproject.toml
[tool.poetry]
name = "asnwerly"
version = "0.1.0"
description = ""
authors = ["DimaK <korbanitskiy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
pre-commit = "^3.3.3"
django-markdownify = "^0.9.3"
django-crispy-forms = "^2.0"
elasticsearch = "7.17.9"
coverage = "^7.2.7"
factory-boy = "^3.2.1"
crispy-bootstrap4 = "^2022.1"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
